apps/backend-api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
alembic==1.14.0
asyncpg==0.29.0
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic-settings==2.6.1
email-validator==2.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
httpx==0.27.2
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0

